ちゃんと直した

+++ .bashrc
@@ -39,6 +39,14 @@ function gnb() {
 alias gmain='git checkout main && git pull'
 
 # あるコミットの状態に完全に戻したいとき
-git reset --hard [コミットID]
-# からの
-git push --force-with-lease
+# git reset --hard [コミットID]
+# git push --force-with-lease
+
+function gre() {
+  if [ -z "$1" ]; then
+    echo "Usage: \"commit sha\""
+    return 1
+  fi
+  git reset --hard "$1"
+  git push --force-with-lease
+}
